app/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.29
python-dotenv==1.0.1
gunicorn==21.2.0
qrcode==7.4.2
Pillow==10.2.0
openpyxl==3.1.2
pytest==8.1.1
pytest-flask==1.3.0
PyJWT==2.8.0
